refactor: Position domain.position로 패키지 이동
