+++ src/main/java/ladder/view/LadderGameController.java
@@ -1,10 +1,10 @@
 package ladder.view;
 
 import ladder.application.LadderGame;
-import ladder.domain.PositionLogs;
 import ladder.domain.line.LadderHeight;
 import ladder.domain.line.LadderLines;
 import ladder.domain.player.Players;
+import ladder.domain.position.PositionLogs;
 import ladder.domain.prize.Prizes;
 
 public class LadderGameController {
