+++ src/test/java/ladder/domain/position/PositionLogFactoryTest.java
@@ -1,7 +1,5 @@
-package ladder;
+package ladder.domain.position;
 
-import ladder.domain.PositionLogFactory;
-import ladder.domain.PositionLogs;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
